pyproject.toml
[tool.poetry]
name = "moqa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain-groq = "^0.1.5"
langchain-huggingface = "^0.0.3"
langchain-community = "^0.2.5"
langchain-chroma = "^0.1.1"
pypdf = "^4.2.0"
python-dotenv = "^1.0.1"
html2text = "^2024.2.26"
tiktoken = "^0.7.0"
beautifulsoup4 = "^4.12.3"
retrying = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
